+++ tuepa/nn/__init__.py
@@ -1,0 +1,1 @@
+from model import *
